--- validator/keymanager/derived/log.go
@@ -1,5 +1,0 @@
-package derived
-
-import "github.com/sirupsen/logrus"
-
-var log = logrus.WithField("prefix", "derived-keymanager")
