+++ validator/keymanager/local/eip_test.go
@@ -1,4 +1,4 @@
-package derived
+package local
 
 import (
 	"encoding/hex"
